Refactor - moving engine to a common static library. Adding a separate console project for unit tests using google's UT framework. Adding tinygltf library for parsing gltf files.

+++ Vulkan_Engine/src/EngineCore/Mesh.h
@@ -18,7 +18,7 @@ struct SubMesh
 	SubMesh() : m_indices(), m_bounds() { }
 	SubMesh(size_t size) : m_indices(), m_bounds() { m_indices.reserve(size); }
 	SubMesh(std::vector<MeshDescriptor::TVertexIndices>& indices) : m_indices(std::move(indices)), m_bounds() { }
-	uint32_t getIndexCount() const { return m_indices.size(); }
+	size_t getIndexCount() const { return m_indices.size(); }
 
 	std::vector<MeshDescriptor::TVertexIndices> m_indices;
 	BoundsAABB m_bounds;
